+++ backend/infrahub/computed_attribute/graphql_queries.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import ClassVar
 
-from infrahub.core.query.node_query import NodeIDQuery
+from infrahub.core.graphql_query.node_query import NodeIDQuery
 
 
 class ComputedAttributeNodeIDQuery(NodeIDQuery):
