+++ backend/infrahub/display_labels/models.py
@@ -1,13 +1,12 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
-from typing import TYPE_CHECKING, Any, ClassVar, Self
+from typing import TYPE_CHECKING, Any, Self
 
 from infrahub_sdk.graphql import Query
 from pydantic import BaseModel, Field
 
 from infrahub.core.constants import RelationshipCardinality
-from infrahub.core.query.node_query import NodeIDQuery
 from infrahub.core.registry import registry
 from infrahub.core.schema import NodeSchema  # noqa: TC001
 from infrahub.events import NodeUpdatedEvent
@@ -162,10 +161,6 @@ class DisplayLabelTriggerDefinition(TriggerBranchDefinition):
             actions=[workflow],
             target_kind=trigger_definition_target_kind,
         )
-
-
-class DisplayLabelNodeIDQuery(NodeIDQuery):
-    query_name: ClassVar[str] = "DisplayLabelFetchNodeIDs"
 
 
 class DisplayLabelJinja2GraphQLResponse(BaseModel):
